Refactoring de materia

+++ src/main/kotlin/Repository/materiasRepository.kt
@@ -3,4 +3,4 @@ package Repository
 import LogicaDeNegocio.Materia
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface materiasRepository : JpaRepository<Materia, Int>  {}
+interface materiasRepository : JpaRepository<Materia, Long>  {}
